Fix potential clock skew in task retry. by suspending on Redis redelivery
When a task fails and there are retries left, it is marked as awaiting_retry. If this state transition succeeds on the server but somehow the worker didn't receive an ack, it will not remove the workflow run id from the stream. This will cause another worker to pick up that workflow run.

At the moment, the new worker checks if there's some delay, and pauses for that duration. This is problematic because the worker's clock might be different from the server's.

Hence, the new worker should simply abandon the workflow if it finds that it's state is `running`. At the right time, the server will schedule the workflow run.

This works cause if the server was responsible for scheduling a workflow run, the run's state will be either `queued`.