State transitions from awaiting states to paused is problematic, hence, it has been disallowed.
Exmaple: suppose a workflow is awaiting an event with a timeout, then it is paused and resumed, on resumption, the timer will be reset. This can cause an infinite loop.
One option to solve this is to keep track of event wait state, similar to how sleep state is tracked, however, this adds more complexity.
Notice that transition from awaiting_event to scheduled is only allowed if the reason is new/event, but allowing pause, opens up a pathway for scheduled reason "resume".

On deeper though I've decided that pause should only be allowed when the worklfow is actively running or is about to be run